requirements.txt
fastapi
h2
httpx
playwright
pydantic
PyYAML
Requests
starlette
uvicorn
